requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
langgraph==0.2.60
langchain==0.3.13
langchain-google-genai==2.0.8
google-generativeai==0.8.3
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
websockets==14.1
